finished separating of controllers, buttons now disappear on retract
