--- src/main/java/se/myhappyplants/client/controller/SearchTabController.java
@@ -1,4 +1,0 @@
-package se.myhappyplants.client.controller;
-
-public class SearchTabController {
-}
