+++ src/main/resources/se/myhappyplants/client/controller/plantsTab.fxml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.text.Font?>
+<?import javafx.geometry.Insets?>
+
+<BorderPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="se.myhappyplants.client.controller.PlantsTabController"
+            prefHeight="400.0" prefWidth="600.0">
+    <top>
+        <AnchorPane prefHeight="56.0" prefWidth="600.0" BorderPane.alignment="CENTER">
+            <children>
+                <Button layoutX="700.0" layoutY="16.0" mnemonicParsing="false" onAction="#logoutButtonPressed" style="-fx-color: Darkred;" text="Log Out" />
+                <ImageView fitHeight="45.0" fitWidth="45.0" layoutX="7.0" layoutY="6.0" pickOnBounds="true" preserveRatio="true" />
+                <Label fx:id="lblUsernamePlants" layoutX="74.0" layoutY="18.0" text="Username">
+                    <font>
+                        <Font size="13.0" />
+                    </font>
+                </Label>
+            </children>
+        </AnchorPane>
+    </top>
+    <center>
+        <Pane prefHeight="200.0" prefWidth="800.0" BorderPane.alignment="CENTER">
+            <children>
+                <TextField fx:id="txtFldSearchText" layoutX="14.0" layoutY="14.0" prefHeight="23.0" prefWidth="441.0" promptText="Search...">
+                    <padding>
+                        <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+                    </padding>
+                </TextField>
+                <Button fx:id="btnSearch" layoutX="467.0" layoutY="15.0" mnemonicParsing="false" onAction="#searchButtonPressed" prefHeight="23.0" text="Search" />
+                <ProgressIndicator fx:id="progressIndicator" layoutX="545.0" layoutY="7.0" progress="0.0" />
+                <ListView fx:id="resultPane" accessibleRole="IMAGE_VIEW" fixedCellSize="-2.0" layoutX="12.0" layoutY="86.0" minHeight="50.0" prefHeight="388.0" prefWidth="778.0" />
+                <Label layoutX="14.0" layoutY="61.0" text="Your results:" />
+            </children>
+        </Pane>
+    </center>
+</BorderPane>
